+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Fixed an issue when viewing `inhibit-read-only`.
 Improved the buffer prompt to be more relevant when inspecting
 buffer-local values.
 
+Better source detection for functions defined by `defstruct` or other
+macros.
+
 # v0.15
 
 Fixed a crash on formatting values.