+++ CHANGELOG.md
@@ -19,6 +19,11 @@ Added looking up C-style Lisp names.
 Set `comment-start` inside helpful buffers, to fix external packages
 relying on that variable.
 
+Helpful now always autoloads callables if they aren't already
+loaded. This is consistent with help.el (unless you've overriden
+`help-enable-auto-load`), produces more useful results, and fixes
+crashes rendering some docstrings.
+
 # v0.15
 
 Fixed a crash on formatting values.